ArM.Advancement

Description

Documentation

advanceSaga :: Saga -> [Saga]

Advance the Saga according to timestamp in the SagaFile.

class Timed a => Advance a where

Advancement

The Advance class represents objects which change state from season to season.

Normally only Saga objects should be advanced explicitly. Other instances exist for internal use in the Saga instance.

Minimal complete definition

nextAdvancement

Methods

nextSeason :: a -> SeasonTime

Next season - if this is undefined (e.g. at GameStart), the time of the next advancement is returned.

nextAdvancement :: a -> SeasonTime

Next season with an advancement defined

prepare :: a -> a

Compute the initial state if no state is recorded. The prepare function is applied when the object is read from file. It is handled differently from in-game advancement, because character generation is independent of any other characters in the game. Thus CharGen advancement never has to infer stats from other objects or check for cross-consistency.

The default implementation is the identity function, which is sufficient for types which starts with a default state and do not suport pre-game advancement.

class StepAdvance c where

StepAdvance is the class of types to which AdvancementStep applies.

Minimal complete definition

nextStep, completeStepMaybe, stepSubjectMaybe

Methods

nextStep :: SeasonTime -> c -> AdvancementStep

completeStep :: AdvancementStep -> c

completeStepMaybe :: AdvancementStep -> Maybe c

stepSubject :: AdvancementStep -> c

stepSubjectMaybe :: AdvancementStep -> Maybe c

data Validation

Validation

A Validation is a message reporting either an error or a successful test.

Constructors

ValidationError String

Validated String

ValidationWarning String