ArM.BasicIO

Description

This module contains convenience functions for file output.

Documentation

data OList

The OList type

Nested lists of strings. This is intended to build output files, where each atomic object is rendered as a list of lines and composite objects as a list of rendered constituent objects.

Constructors

OList [OList]

OString String

toOList :: [String] -> OList

Convert a list of Strings to a OList object

foldOList :: OList -> OList

indentOList :: OList -> OList

Render an OList as a hierarchical markdown list

indentOList' :: String -> OList -> OList

indentOList1 :: OList -> OList

Render an OList as a hierarchical markdown list

indentOList1' :: String -> OList -> OList

writeOListH :: Handle -> OList -> IO ()

Writing OList to file

writeOList :: String -> OList -> IO ()

writeMaybeOList :: Maybe String -> OList -> IO ()

writeLns

Arguments

:: String
File name

-> [String]
Contents

-> IO ()

File output

Write a list of strings into the given file

writeMaybeFile

Arguments

:: Maybe String
File name

-> [String]
Contents

-> IO ()

writeMaybeJSON :: ToJSON t => Maybe String -> t -> IO ()

putStrLns :: [String] -> IO ()