ArM.Char.Advancement

Description

Documentation

module ArM.Types.Advancement

module ArM.Char.Virtues

module ArM.Char.Inference

prepareAdvancement :: CharacterState -> Advancement -> AugmentedAdvancement

Preparing the Advancement

Augment and amend the advancements based on current virtues and flaws.

applyAdvancement :: AugmentedAdvancement -> CharacterState -> (AugmentedAdvancement, CharacterState)

Applying the Advancement

Apply advancement This function is generic, and used for both chargen and ingame advancement. The AugmentedAdvancement has to be prepared differently, using either prepareAdvancement or prepareCharGen.

agePT

Arguments

:: Int
Number of years

-> ProtoTrait
Resulting ProtoTrait

Convenience Functions (Exported)

Return a ProtoTrait for aging advancing a number of years.