ArM.Char.CharGen

Description

The module exports only one function, prepareCharacter, which computes the character at game start, based on pre-game advancement objects.

This function should be applied when a character is read from file which does not include a persistent state.

Documentation

prepareCharacter :: Character -> Character

Char Gen

Compute the initial state if no state is recorded. The function uses applyCGA to process all of the pregame advancements. It then calls addConfidence to add the confidence trait to the state for the returned Character object