ArM.Char.Character

Description

This module contains exports the types to process characters and advancement, including persistence in JSON and advancement.

Documentation

module ArM.Char.Types.Character

data KeyPairList

Constructors

KeyPairList [KeyPair]

data KeyPair

Constructors

KeyPair

Fields

key :: String

value :: FieldValue

data FieldValue

Constructors

TextValue Text

IntValue Int

ObjectValue Value

characterStateName :: Character -> String

Character name with state identifier (current season)

characterStartName :: Character -> String

Character name with state identifier at Game Start

parseSeasonTime :: Maybe String -> SeasonTime

data Advancement

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types.

Constructors

Advancement

Fields

advMode :: Maybe String

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: Maybe String

freeform description of the activities

advUses :: Maybe [Resource]

Books and other resources used exclusively by the character

advSQ :: Maybe XPType

Source Quality (SQ) This should be the common SQ for adventures; individual variation should be recorded as advBonus

advBonus :: Maybe XPType

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

data AugmentedAdvancement

Advancement with additional inferred fields

Constructors

Adv

Fields

advancement :: Advancement

Base advancement as entered by the user

effectiveSQ :: Maybe XPType

SQ modified by virtues and flaws

levelLimit :: Maybe Int

spell level allowance

spentXP :: Maybe XPType

Total XP spent on advancement

inferredTraits :: [ProtoTrait]

trait changes inferred by virtues and flaws

augYears :: Maybe Int

number of years advanced

validation :: [Validation]

Report from validation

class AdvancementLike a where

Methods

mode

Arguments

:: a

-> Maybe String
mode of study

season

Arguments

:: a

-> SeasonTime
season or development stage

narrative

Arguments

:: a

-> Maybe String
freeform description of the activities

uses

Arguments

:: a

-> Maybe [Resource]
Books and other resources used exclusively by the character

sourceQuality

Arguments

:: a

-> Maybe XPType
Source Quality (SQ) effectiveSQ :: Maybe Int -- ^ SQ modified by virtues and flaws

changes

Arguments

:: a

-> [ProtoTrait]
trait changes defined by player inferredTraits :: [ ProtoTrait ] -- ^ trait changes inferred by virtues and flaws

prepareCharacter :: Character -> Character

Char Gen

Compute the initial state if no state is recorded.

The function uses applyCGA to process all of the pregame advancements. It then calls addConfidence to add the confidence trait to the state for the returned Character object

class Advance a where

Methods

advance :: SeasonTime -> a -> a

Advance the character until after the given time.

step :: a -> a

Advance the character one season forward

nextSeason :: a -> SeasonTime

Time of the next advancement of the character.

data Season

Season of the year. This is currently defined using the Hibernian calendar, with Winter being the first season of the year. This must be changed for the standard calendare. data Season = Spring | Summer | Autumn | Winter | NoSeason

Constructors

Winter

Spring

Summer

Autumn

NoSeason

data SeasonTime

Constructors

SeasonTime Season Int

GameStart

NoTime

characterEntryTime :: Character -> SeasonTime

The first season the character is played

isGameStart :: Character -> Bool

The first season the character is played