ArM.Char.CharacterSheet

Description

The CharacterSheet type is a façade exposing lists for each kind of trait. The module also includes convenience functions to calculate derived stats such as casting totals.

Documentation

data CharacterSheet

The CharacterSheet object holds a Character State with separate fields for different kinds of traits

Constructors

CharacterSheet

Fields

csType :: CharacterType

vfList :: [VF]

charList :: [Characteristic]

abilityList :: [Ability]

artList :: [Art]

spellList :: [Spell]

reputationList :: [Reputation]

ptList :: [PTrait]

confList :: [Confidence]

otherList :: [OtherTrait]

possessionList :: [Possession]

combatList :: [CombatOption]

csTraits :: [Trait]

characterSheet :: Character -> CharacterSheet

Get the CharacterSheet from a given Character.

filterCS :: CharacterState -> CharacterSheet

Get the CharacterSheet corresponding to a given CharacterState.

sheetArtScore :: CharacterSheet -> TraitKey -> Int

Get a given art score

sheetAbilityScore :: CharacterSheet -> TraitKey -> (Int, Maybe String)

Get the score and speciality in a given ability.

sheetCharacteristicScore :: CharacterSheet -> TraitKey -> Int

Get a given characteristic score

sheetPossession :: CharacterSheet -> TraitKey -> Maybe Possession

Get a given Possession from the character sheet

castingScore

Arguments

:: SpellDB
Spell DB with general descriptions of the spells

-> CharacterSheet
Current character sheet

-> TraitKey
Key identifying the spell

-> Int
Computed casting score

Return the Casting Score for a given spell. The function depends both on the Spell trait from the CharacterSheet and a generic spell description from a SpellDB.

addCastingScores :: SpellDB -> CharacterSheet -> CharacterSheet

labTotals

Arguments

:: CharacterSheet
Current character sheet

-> [[Int]]
Computed lab totals

class HasAge a where

Character Age

Methods

age :: a -> Int

ageObject :: a -> Maybe Age

sheetVis :: CharacterSheet -> [(TraitKey, Int)]