ArM.Char.Combat

Description

This module proves a type for each kind of trait as well as a wrapper type, Trait which can represent any kind of trait.

This module defines the types as well as the TraitType class, and instances for show, sorting, and JSON.

Documentation

data CombatLine

The Trait Type

Constructors

CombatLine

Fields

combatLabel :: String

combatInit :: Int

combatAtk :: Maybe Int

combatDef :: Maybe Int

combatDam :: Maybe Int

combatRange :: Maybe Int

combatLoad :: Int

combatComment :: String

defaultCL :: CombatLine

data SoakLine

Constructors

SoakLine

Fields

soakLabel :: String

soakScore :: Int

soakLoad :: Int

computeCombatLine :: WeaponDB -> CharacterSheet -> CombatOption -> CombatLine

Compute a single line for the Combat Stats table

addCharacteristics :: CharacterSheet -> CombatLine -> CombatLine

Add characteristics to the combat stats

weaponStat :: WeaponDB -> Possession -> [Weapon]

Look up weapon stats from a possession Standard stats from the DB and specific stats from the item are concatenated.

sheetWeapon :: WeaponDB -> CharacterSheet -> String -> [Weapon]

Look up weapon stats from a character sheets. If a possession is not found, it is looked up in the WeaponDB instead

sheetWeapon2 :: WeaponDB -> String -> [Weapon]

sheetWeapon1 :: WeaponDB -> CharacterSheet -> String -> [Weapon]

implicitAbility :: WeaponDB -> CharacterSheet -> CombatOption -> CombatLine -> CombatLine

Compute the CombatLine from a CombatOption without explicit ability

addShield :: WeaponDB -> CharacterSheet -> String -> Maybe String -> CombatLine -> CombatLine

explicitAbility :: WeaponDB -> CharacterSheet -> CombatOption -> String -> CombatLine -> CombatLine

Compute the CombatLine from a CombatOption with explicit ability

join :: Maybe (Maybe a) -> Maybe a

Collapse a neste Maybe Maybe object to a single Maybe

computeCombatLines :: WeaponDB -> CharacterSheet -> [CombatOption] -> [CombatLine]

Compute the table of combat stas from a list of CombatOption objects

computeCombatStats :: WeaponDB -> CharacterSheet -> [CombatLine]

Compute the table of combat stas from a list of CharacterSheet objects