ArM.Char.Inference

Description

Only one function is exported: addInference which is used for both in-game and chargen advancement. It must be used before all other processing of the advancement, since it transforms an Advancement to an AugmentedAdvancement.

Documentation

addInference :: CharacterState -> Advancement -> AugmentedAdvancement

Infer traits a range of other traits, both from the new advancement and the existing CharacterState.