ArM.Char.Types.Character

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data Character

Character

The Character object includes both state information and timeless concept information, as well as the advancements defining the evolution through states.

Constructors

Character

Fields

charID :: String

character ID, used to cross-reference

concept :: CharacterConcept

concept is the timeless features of the character

state :: Maybe CharacterState

current state of the character

entryTime :: SeasonTime

First season the character is in play

pregameDesign :: [AugmentedAdvancement]

chargen already processed

pregameAdvancement :: [Advancement]

chargen left to process

pastAdvancement :: [AugmentedAdvancement]

past advancement (in game), most recent first

futureAdvancement :: [Advancement]

future advancement (in game), next one firstk

defaultCharacter :: Character

Default (empty) character object.

data CharacterID

Name and identity

ID of a character. This is currently implemented as the name. The ID is used to refer to the character from other characters and covenants, independently of the current state.

Constructors

CharacterID String

characterID :: Character -> CharacterID

get the ID of a character.

fullName :: Character -> String

Return the name of the character as a string, including house affiliation if defined.

fullConceptName :: CharacterConcept -> String

Return the name of the character as a string, including house affiliation if defined.

characterSeason :: Character -> SeasonTime

data CharacterType

CharacterConcept

The CharacterType distinguishes between Magus, Companion, and Grog. One may want to extend it for magic and faerie charaacters as well as NPCs.

Constructors

Magus

Companion

Grog

class CharacterLike ct where

Minimal complete definition

characterType

Methods

characterType :: ct -> CharacterType

isGrog :: ct -> Bool

Is the character a grog or not?

isMagus :: ct -> Bool

Is the character a magus or not?

data CharacterConcept

The CharacterConcept is the timeless information about the character.

Constructors

CharacterConcept

Fields

name :: String

charType :: CharacterType

briefConcept :: Maybe String

quirk :: Maybe String

appearance :: Maybe String

born :: Maybe Int

player :: Maybe String

house :: Maybe String

charGlance :: KeyPairList

charData :: KeyPairList

defaultConcept :: CharacterConcept

Default (empty) character concept object.

data CharacterState

CharacterState

The Character state is the stats of the character at a particular point in time.

Constructors

CharacterState

Fields

charTime :: SeasonTime

charSType :: CharacterType

traits :: [Trait]

defaultCS :: CharacterState

Default (empty) character state object.