ArM.Char.Types.Trait

Description

This module proves a type for each kind of trait as well as a wrapper type, Trait which can represent any kind of trait.

This module defines the types as well as the TraitType class, and instances for show, sorting, and JSON.

Documentation

data TraitKey

A unique identifier for traits. It implement the Ord class, with canonical ordering on characteristics and arts.

Constructors

AbilityKey String

CharacteristicKey String

ArtKey String

SpellKey String Int String

PTraitKey String

ReputationKey String String

VFKey String String

ConfidenceKey String

OtherTraitKey String

SpecialKey String

PossessionKey String

CombatKey String

AgeKey

NoTrait

data Trait

The Trait Type

Constructors

AbilityTrait Ability

CharacteristicTrait Characteristic

ArtTrait Art

SpellTrait Spell

PTraitTrait PTrait

ReputationTrait Reputation

VFTrait VF

ConfidenceTrait Confidence

OtherTraitTrait OtherTrait

SpecialTraitTrait SpecialTrait

PossessionTrait Possession

CombatOptionTrait CombatOption

AgeTrait Age

data Ability

Different types of Traits

Constructors

Ability

Fields

abilityName :: String

speciality :: Maybe String

abilityXP :: XPType

abilityScore :: Int

abilityBonus :: Int

abilityMultiplier :: Float

abilityExcessXP :: XPType

data Characteristic

Constructors

Characteristic

Fields

characteristicName :: String

charScore :: Int

agingPoints :: Int

charBonusList :: [(Int, Int)]

data Art

Constructors

Art

Fields

artName :: String

artXP :: XPType

artScore :: Int

artBonus :: Int

artMultiplier :: Float

artExcessXP :: XPType

data VF

Constructors

VF

Fields

vfname :: String

name of the virtue/flaw

vfDetail :: String

detail, where the virtue/flaw has options

vfcost :: Int

cost, should be zero for freeinferred virtuesflaws

vfAppliesTo :: Maybe TraitKey

not used

vfMultiplicity :: Int

number of times the virtue/flaw is take

vfComment :: String

freeform comment

data Spell

Constructors

Spell

Fields

spellName :: String

spellTeFo :: String

spellLevel :: Int

spellXP :: XPType

masteryScore :: Int

spellExcessXP :: XPType

spellMultiplier :: Float

masteryOptions :: [String]

spellCastingScore :: Maybe Int

spellTComment :: String

data PTrait

Personality trait

Constructors

PTrait

Fields

ptraitName :: String

pscore :: Int

data SpecialTrait

Constructors

SpecialTrait

Fields

specialTrait :: String

specialScore :: Maybe Int

specialComment :: Maybe String

data OtherTrait

Constructors

OtherTrait

Fields

trait :: String

otherScore :: Int

otherExcess :: Int

data Possession

Weapons and other Possessions

A Possession is any kind of device that can be acquired, lost, given, or traded. It is treated like inherent traits in the data model. Possessions comprise weapons, armour, vis, magic devices, equipment, and any physical object that should be recorded on the characters sheet.

Constructors

Possession

Fields

itemName :: String

Name identifying the unique item

weaponStats :: [Weapon]

List of applicable Weapon stat objects

weapon :: [String]

List of standard weapon stats that apply

armourStats :: [Armour]

List of applicable Weapon stat objects

armour :: [String]

List of standard weapon stats that apply

itemDescription :: String

Description of the Item

visArt :: Maybe String

Relevant art if the item is raw vis

itemCount :: Int

Number of items possessed, default 1.

data CombatOption

Combat Options

A CombatOption is a combination of weapons for which to list combat stats.

It needs to link with a Weapon (and optionally a shield) which may be generic or unique. If the weapon can be used in different modes, the ability has to be linked as well.

Constructors

CombatOption

Fields

combatName :: String

Describing Name of the weapon combination

combatWeapon :: String

The main weapon

combatShield :: Maybe String

A Shield is optional

combatAbility :: Maybe String

data Confidence

Constructors

Confidence

Fields

cname :: String

cscore :: Int

cpoints :: Int

data Reputation

Reputation object

Constructors

Reputation

Fields

reputationName :: String

contents of the reputation

repLocale :: String

domain or location of the reputation

repXP :: XPType

total XP in the reputation (used?)

repScore :: Int

reputation Score

repExcessXP :: XPType

XP towards next level in the reputation

data Age

Constructors

Age

Fields

ageYears :: Int

character age in years

apparentYounger :: Int

difference between age and apparent age

ageLimit :: Int

longevityRitual :: Int

Score of longevity ritual (LR), negative number means none

agingRollBonus :: Int

Bonus to aging rolls (excluding LR)

ageComment :: Maybe String

freeform comment

class TraitClass t where

TraitClass

TraitClass provides the functions to get the search key (TraitKey), to wrap and unwrap traits in the generic Trait type, and to filter traits of different types.

ProtoTrait and its constituent types may also implement TraitClass but getTrait may then always return Nothing.

Minimal complete definition

traitKey, toTrait, getTrait

Methods

traitKey :: t -> TraitKey

Get the key of the trait

toTrait :: t -> Trait

Wrap the trait as a generic Trait object.

getTrait :: Trait -> Maybe t

Return the specific trait from the generic Trait, or Nothing if the type does not match.

filterTrait :: [Trait] -> ([t], [Trait])

Extract traits of the given type from a generic list of Trait objects. It returns a pair of lists with the selected traits in the first list and the remaining traits in the other.

filterTrait' :: ([Trait], ([t], [Trait])) -> ([Trait], ([t], [Trait]))

Recursive helper for filterTrait

(<:) :: (TraitClass t1, TraitClass t2) => t1 -> t2 -> Bool

Sorting

fote :: String -> String

Convert the TeFo string to Form/Technique for sorting

sortTraits :: TraitClass t => [t] -> [t]