ArM.Char.Validation

Description

Only one function is exported: validate. It checks the integrity of an AugmentedAdvancement object, and adds its reports to the validation field.

This function should be used after all other calculation of the AugmentedAdvancement is complete, and before the object is displayed, or written to file.

Documentation

validate :: AugmentedAdvancement -> AugmentedAdvancement

In-game Validation

In-game validation is relatively simple, depending only on the AugmentedAdvancement. Currently, only XP expenditure is validated.

Validate an in-game advancement, adding results to the validation field.

validateCharGen :: CharacterSheet -> AugmentedAdvancement -> AugmentedAdvancement

CharGen Validation

CharGen validation is tricky, often depending on virtues and flaws. Therefore, most functions depend also on the CharacterSheet in addition to the AugmentedAdvancement.

validate an advancement, adding results to the validation field

initialLimits :: CharacterSheet -> AugmentedAdvancement -> AugmentedAdvancement

Calculate initial XP limits on Char Gen Advancements