ArM.Cov.Covenant

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data Covenant

Covenant Object

A Covenant consists of a state and a timeless concept, as well as lists of advancement which define the evolution of states

Constructors

Covenant

Fields

covenantConcept :: CovenantConcept

covenantState :: Maybe CovenantState

pastCovAdvancement :: [AugCovAdvancement]

futureCovAdvancement :: [CovAdvancement]

data CovenantID

ID of a Covenant. This is currently implemented as the name. It is used to reference the covenant, independtly of state, from other objects.

Constructors

CovenantID String

covenantID :: Covenant -> CovenantID

get the ID of a character.

covenant :: CharacterState -> Maybe CovenantID

data CovenantConcept

CovenantConcept Object

Constructors

CovenantConcept

Fields

covName :: String

covConcept :: Maybe String

covFounded :: Maybe Int

covAppearance :: Maybe String

covTribunal :: Maybe String

covData :: KeyPairList

defaultCovConcept :: CovenantConcept

Default (empty) covenant concept object.

data CovenantState

CovenantState Object

Constructors

CovenantState

Fields

covTime :: SeasonTime

covenFolkID :: [CharacterID]

library :: [Book]

librarycsv :: Maybe String

defaultCovState :: CovenantState

data CovAdvancement

Advancement and Traits

Advancement (changes) to a covenant.

Constructors

CovAdvancement

Fields

caSeason :: SeasonTime

season or development stage

caNarrative :: String

freeform description of the activities

joining :: [CharacterID]

leaving :: [CharacterID]

acquired :: [Book]

lost :: [Book]

defaultAdv :: CovAdvancement

data AugCovAdvancement

Augmented advancement for covenants. This comprises changes inferred from the advancements of characters (and possibly other covenants) as well as the advancement from the input file.

Constructors

AugCovAdvancement

Fields

explicit :: Maybe CovAdvancement

inferred :: Maybe CovAdvancement

joiningAug :: AugCovAdvancement -> [CharacterID]

CovenFolk joining according to the augmented covenant advancement.

leavingAug :: AugCovAdvancement -> [CharacterID]

CovenFolk leaving according to the augmented covenant advancement.

caSeasonAug :: AugCovAdvancement -> SeasonTime

Get the season of the augmented covenant advancement. This is taken from the explicit advancement if available, and the inferred advancement otherwise.

applyCovAdv :: (Covenant, Maybe AugCovAdvancement) -> (Covenant, Maybe AugCovAdvancement)

Apply covenant advancement

findCov :: Character -> [Covenant] -> Maybe Covenant

Find the character's covenant from a list. The covenant is identified by checking if the character is listed as a member (covenFolkID).

findBook :: CovenantState -> String -> Maybe Book

hasMember :: Covenant -> Character -> Bool

Does the covenant have the character as a member?

completeCovAdv :: (Covenant, Maybe AugCovAdvancement) -> Covenant

Complete the advancement procedure to return the new Covenant with the updated state.

nextCovAdv :: SeasonTime -> Covenant -> (Covenant, Maybe AugCovAdvancement)

Get the next augmented advancement.

prepareAdvancement :: CovAdvancement -> AugCovAdvancement