ArM.DB.Spell

Description

Documentation

data SpellRecord

A SpellRecord is a shared object describing a spell. It is different from the Spell trait which represents the individual's knowledge of the spell. Note that the SpellRecord is identified by spell name, so that generic level spells share a record. The SpellKey is used to sort traits and includes the level of the instance.

Constructors

SpellRecord

Fields

spellRecordName :: String

Name of the spell

spellRecordTeFo :: String

Technique/Form abreviation for the spell

lvl :: Maybe Int

Spell Level. General Level Spells have Nothing.

technique :: String

techniqueReq :: [String]

form :: String

formReq :: [String]

rdt :: (String, String, String)

RangeDurationTarget

specialSpell :: [String]

Special tags, like Ritual or Mutantum.

description :: String

Freeform description of the effect

design :: String

Level calculation

spellComment :: String

Freeform remarks that do not fit elsewhere

cite :: String

Source reference

spellDB :: [[String]] -> SpellDB

Create a Map of SpellRecord objects. The input is the output from csvFile

spellLookup :: TraitKey -> SpellDB -> Maybe SpellRecord

type SpellDB = Map String SpellRecord

Default SpellRecord object as a starting point for step-by-step construction.