ArM.Debug.NoTrace

Description

This module defines the same functions as Trace to be identity functions; thus the debug trace can be turned off simply by importing this module instead of Trace, leaving the debug code in place.

Documentation

trace :: String -> a -> a

ttrace :: Show a => a -> a

strace :: String -> String