ArM.IO

Description

The functions in this module generally take a filename (String) and returns a Maybe object by attempting to parse the relevant file.

Some objects, like the Spell Database, are stored in CSV, others in JSON, including Character and Covenant.

Documentation

readSagaFile

Arguments

:: String
Filename

-> IO (Maybe SagaFile)

Read Saga Files

| Read a saga from JSON. Return Maybe SagaFile

readSaga

Arguments

:: String
Filename

-> IO (Maybe Saga)

Load the saga and its constituent objects from the given file. This calls both readSagaFile and loadSaga.

loadSaga :: SagaFile -> IO Saga

Load constituent objects for a saga.

writeSagaState :: Saga -> SagaState -> IO ()

writeSagaStates :: Saga -> [SagaState] -> IO ()

writeSaga :: Saga -> IO ()

class FromJSON c => ArMRead c where

Read Character and Covenant Data

Minimal complete definition

prepare

Methods

readArM

Arguments

:: String
Filename

-> IO (Maybe c)

Read a character from JSON. Return Maybe Character

prepare :: c -> c

class LongSheet h => HarmObject h where

Write Character Sheets

Minimal complete definition

stateName

Methods

writeObjects

Arguments

:: String
Directory for the output files

-> Saga
Saga whose objects are written

-> [h]
List of objects to write

-> IO ()

Write charactersheets at Game Start in MarkDown File name is derived from the character name.

stateName :: h -> String