ArM.Markdown

Description

The core of this module is the Markdown class and its printMD function which renders an object in Markdown. There is also a LongSheet class with a printSheetMD function for a more verbose character sheet.

Documentation

class Markdown a where

The Markdown Class

Class defining printMD to render in Markdown.

Minimal complete definition

printMD

Methods

printMD

Arguments

:: a
object to render

-> OList
list of lines for output

This is the basic function to render in Markdown

printMDaug

Arguments

:: Saga
Saga including databases for spells etc.

-> a
object to render

-> OList
list of lines for output

This is a hack to augment characters using extra resources By default, it is identical to printMD.

class Markdown a => LongSheet a where

Long Sheet Format

The LongSheet class is similar to Markdown, but is intended to make a longer output with both more space and more verbose text.

In the current impplementation, it is only the character state which is made longer, by setting abilities and spells as bullet points instead of a single paragraph for the full list.

Minimal complete definition

Nothing

Methods

printSheetMD

Arguments

:: Saga
Saga including databases for spells, etc.

-> a
object to render

-> OList
list of lines for output

By default printSheetMD is identical to printMDaug

gameStartSheet :: Saga -> Character -> OList

Render a character sheet at game start. Unlike the regular printMD, this includes only the character design and not ingame advancements.

currentSheet :: Saga -> Character -> OList

Render the current character sheet without pregame design details.

artMD :: CharacterSheet -> OList

Pretty print arts

Render art scores as a table

artVisMD :: CharacterSheet -> OList

Render art scores and vis stocks as a table