ArM.Types.Advancement

Description

Advancement is tricky because there are many different modes of study and many special cases. The basic structure uses two main types: + Advancement is the changes defined by the user. + AugmentedAdvancement comprises the Advancement object and additional inferred changes.

The Source Quality (SQ) is controled by several fields. + AugmentedAdvancement baseSQ is the base source quality inferred from other traits and general rules. + Advancement sourceQuality allows the user to enter the basic SQ manually. This is required for Practice, and may be used to override the baseSQ. If it differs from baseSQ, a warning is issued. + AugmentedAdvancement bonusSQ includes modifications derived automatically from virtues and flaws, and other individual circumstances. + Advancement bonusQuality should be used for individual modifications that cannot be automatically inferred, such as correspondent and Study Bonus.

Also not the scoreCap or sourceCap that caps the score that can be acquired from the source. This applies to books, trainers, and teachers.

The Advancement object has a list of changes which is ProtoTrait objects modifying existing traits. Similarly, AugmentedAdvancement has inferredTraits for additional implied changes.

Documentation

data AdvancementType

Advancement

Advancement Types

Main advancement modes as defined in the core rules [ArM]

Constructors

Practice

Adventure

Taught

Trained

Reading

VisStudy

Exposure ExposureType

CharGen String

data ExposureType

Different activities that grant Exposure but may require different processing in advancement and validation.

Constructors

LabWork

Teaching

Training

Writing

Copying

Authoring

Initiation

OpeningArts

Work

SpellInstruction

OtherExposure String

parseET :: String -> ExposureType

Parse an ExposureType from a String, auxiliary for FromJSON

parseAT :: String -> AdvancementType

Parse an AdvancementType from a String, auxiliary for FromJSON

dropWord :: String -> String

Drop the first word from a string.

class StoryObject a => AdvancementLike a where

AdvancementLike Class

The AdvancementLike class gives a common API to Advancement and AugmentedAdvanceemnt

Minimal complete definition

mode, usesBook, sourceQuality, sourceCap, changes

Methods

mode

Arguments

:: a

-> AdvancementType
mode of study

usesBook

Arguments

:: a

-> [String]
Books used exclusively by the character

sourceQuality

Arguments

:: a

-> Maybe XPType
Source Quality (SQ)

sourceCap

Arguments

:: a

-> Maybe Int
Level cap from the source of learning

changes

Arguments

:: a

-> [ProtoTrait]
trait changes defined by player

isExposure :: a -> Bool

data Advancement

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types. Note that standard SQ should be recorded as advSQ, while individual variation may be recorded as advBonus.

Constructors

Advancement

Fields

advMode :: AdvancementType

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: [String]

narrative description of the activities

advComment :: [String]

freeform description of the activities

advUses :: [String]

Books used exclusively by the character , advReadTome :: String -- ^ Tome studied , advReadBook :: String -- ^ Partook studied , advReadTrait :: String -- ^ Book studied

advSQ :: Maybe XPType

Source Quality (SQ)

advCap :: Maybe Int

Source Quality (SQ)

advBonus :: Maybe XPType

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

defaultAdv :: Advancement

data AugmentedAdvancement

The Augmented Advancement

Advancement with additional inferred fields

Constructors

Adv

Fields

advancement :: Advancement

Base advancement as entered by the user

baseSQ :: Maybe XPType

Base Source Quality

bonusSQ :: XPType

Bonus to Source Quality from Virtues and Flaws

scoreCap :: Maybe Int

levelLimit :: Maybe Int

spell level allowance

spentXP :: Maybe XPType

Total XP spent on advancement

inferredTraits :: [ProtoTrait]

trait changes inferred by virtues and flaws

augYears :: Maybe Int

number of years advanced

validation :: [Validation]

Report from validation

postProcessTrait :: PostProcessor

Extra postprocessing for traits at the given stage

bookUsed :: [Book]

Books required by the activity

teacherSQ :: Maybe Int

data PostProcessor

Type of function used to post-process traits after advancement.

Constructors

PostProcessor (Trait -> Trait)

defaultAA :: AugmentedAdvancement

showTime :: String -> SeasonTime -> AdvancementType -> Maybe Int -> String

Render the season and mode of an advancement

showYears :: Maybe Int -> String

Render the duration of an advancement

showSQ :: Maybe XPType -> XPType -> String

Render the source quality of an advancement

effectiveSQ :: AugmentedAdvancement -> XPType

data Validation

Validation

A Validation is a message reporting either an error or a successful test.

Constructors

ValidationError String

Validated String

primaryXPTrait :: Advancement -> Maybe TraitKey

sortInferredTraits :: AugmentedAdvancement -> AugmentedAdvancement

Sort the inferredTraits field of an AugmentedAdvancement.