ArM.Types.Character

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data Character

Character

The Character object includes both state information and timeless concept information, as well as the advancements defining the evolution through states.

Constructors

Character

Fields

charID :: String

character ID, used to cross-reference

concept :: CharacterConcept

concept is the timeless features of the character

state :: Maybe CharacterState

current state of the character

entryTime :: SeasonTime

First season the character is in play

pregameDesign :: [AugmentedAdvancement]

chargen already processed

pregameAdvancement :: [Advancement]

chargen left to process

pastAdvancement :: [AugmentedAdvancement]

past advancement (in game), most recent first

futureAdvancement :: [Advancement]

future advancement (in game), next one firstk

defaultCharacter :: Character

Default (empty) character object.

fullConceptName :: CharacterConcept -> String

Return the name of the character as a string, including house affiliation if defined.

characterSeason :: Character -> SeasonTime

data CharacterType

CharacterConcept

The CharacterType distinguishes between Magus, Companion, and Grog. One may want to extend it for magic and faerie charaacters as well as NPCs.

Constructors

Magus

Companion

Grog

data CharacterConcept

The CharacterConcept is the timeless information about the character.

Constructors

CharacterConcept

Fields

charName :: String

charType :: CharacterType

briefConcept :: Maybe String

quirk :: Maybe String

appearance :: Maybe String

born :: Maybe Int

Year of Birth

player :: Maybe String

player responsible for the character

house :: Maybe String

Hermetic House or similar affiliation

portrait :: Maybe String

image file name (portrait)

charGlance :: KeyPairList

charData :: KeyPairList

defaultConcept :: CharacterConcept

Default (empty) character concept object.

data CharacterState

CharacterState

The Character state is the stats of the character at a particular point in time.

Constructors

CharacterState

Fields

charTime :: SeasonTime

charSType :: CharacterType

memberOf :: Maybe String

traits :: [Trait]

defaultCS :: CharacterState

Default (empty) character state object.