ArM.Types.Covenant

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data Covenant

Covenant Object

A Covenant consists of a state and a timeless concept, as well as lists of advancement which define the evolution of states

Constructors

Covenant

Fields

covenantConcept :: CovenantConcept

covenantState :: Maybe CovenantState

pastCovAdvancement :: [AugCovAdvancement]

futureCovAdvancement :: [CovAdvancement]

data CovenantConcept

CovenantConcept Object

Constructors

CovenantConcept

Fields

covName :: String

covConcept :: Maybe String

covDescription :: [String]

covFounded :: Maybe Int

covAppearance :: Maybe String

covTribunal :: Maybe String

covData :: KeyPairList

data CovenantState

CovenantState Object

Constructors

CovenantState

Fields

covTime :: SeasonTime

covenFolkID :: [HarmKey]

library :: [Book]

librarycsv :: Maybe String

labs :: [Lab]

data CovAdvancement

Advancement and Traits

Advancement (changes) to a covenant.

Constructors

CovAdvancement

Fields

caSeason :: SeasonTime

season or development stage

caStory :: [Story]

freeform description of the activities

joining :: [HarmKey]

leaving :: [HarmKey]

acquired :: [Book]

lost :: [Book]

data AugCovAdvancement

Augmented advancement for covenants. This comprises changes inferred from the advancements of characters (and possibly other covenants) as well as the advancement from the input file.

Constructors

AugCovAdvancement

Fields

explicit :: Maybe CovAdvancement

inferred :: Maybe CovAdvancement

findCov :: Character -> [Covenant] -> Maybe Covenant

Find the character's covenant from a list. The covenant is identified by checking if the character is listed as a member (covenFolkID).

defaultCovState :: CovenantState

findBook :: CovenantState -> String -> Maybe Book

contractAdvancement :: AugCovAdvancement -> CovAdvancement

Merge explicit and inferred advancement into onw CovAdvancement object

covenant :: CharacterState -> Maybe HarmKey