ArM.Types.HarmObject

Description

HarmKey type used to index arbitrary hArM objects.

Documentation

data HarmKey

A unique identifier for objects. It is made quite generic to support a class KeyObject of keyed objects without enabling multi-parameter classes.

Constructors

BookKey String

CharacterKey String

CovenantKey String

LabKey String

NoObject

class (Timed h, StoryObject h) => HarmObject h where

The HarmObject class establishes a common interface for Covenant and Character.

Minimal complete definition

Nothing

Methods

stateName :: h -> String

String identifying the object and its state

isGameStart :: h -> Bool

Is the character state still at Game Start?

class KeyObject h where

Class of types that can be indexed by HarmKey objects.

Minimal complete definition

harmKey

Methods

harmKey :: h -> HarmKey

Return the unique key of the object

harmFind :: HarmKey -> [h] -> Maybe h

Find an object by key in a list.

harmFilter :: HarmKey -> [h] -> [h]

Find an object by key in a list.

harmLookup :: [HarmKey] -> [h] -> [h]

Find a list of objects by key in a sorted list.

sortOnKey :: [h] -> [h]

Sorty objects by HarmKey

class StoryObject ob where

Common interface for objects that have a narrative aspect.

Minimal complete definition

name

Methods

name :: ob -> String

The name could be the title of the story or other unique identifier.

narrative :: ob -> [String]

Description focusing on a narrative feel.

comment :: ob -> [String]

Remarks that do not fit in a narrative style, e.g. mechanics.

class Countable c where

Class for countable objects.

Methods

count :: c -> Int

addCount :: c -> Int -> c

compareKey :: (KeyObject a, KeyObject b) => a -> b -> Ordering

Compare to objects by their HarmKey.