ArM.Types.KeyPair

Description

The KeyPairList type is created to represent arbitrary objects parsed from JSON. It is not necessarily completely generic, but supports a generic list of label/value pairs, where the value may be a String or an Int. This supports custom fields in character sheets.

Documentation

data FieldValue

A FieldValue is the type of the value in a KeyPair, supporting Text or Int, or a catch-all JSON value. The catch-all Value case prevents some parsing errors from JSON.

Constructors

TextValue Text

IntValue Int

ObjectValue Value

data KeyPair

A KeyPair is one label/value pair.

Constructors

KeyPair

Fields

key :: String

value :: FieldValue

data KeyPairList

A KeyPairList is a list of label/value pairs.

Constructors

KeyPairList [KeyPair]