ArM.Types.Library

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data BookOriginal

Constructors

BookOriginal

Fields

bookTitle :: String

bookStats :: [BookStats]

bookAuthor :: String

bookYear :: Int

bookLocation :: String

bookAnnotation :: String

data BookKey

Type for the unique identifier of an original book

Constructors

BookKey String

bookKey :: BookOriginal -> BookKey

Get the unique identifier of an original book

data BookCopy

Constructors

BookCopy

Fields

copyTitle :: String

copyStats :: [BookStats]

copyist :: String

copyYear :: Int

copyLocation :: String

copyAnnotation :: String

original :: BookKey

data BookStats

Constructors

BookStats

Fields

topic :: TraitKey

quality :: Int

bookLevel :: Maybe Int