ArM.Types.Story

Description

Documentation

data Story

Constructors

Story

Fields

storySeason :: SeasonTime

storyTitle :: String

storyNarrative :: [String]

storyComment :: [String]

storySQ :: Maybe Int

class StoryObject ob where

Common interface for objects that have a narrative aspect.

Minimal complete definition

name

Methods

name :: ob -> String

The name could be the title of the story or other unique identifier.

narrative :: ob -> [String]

Description focusing on a narrative feel.

comment :: ob -> [String]

Remarks that do not fit in a narrative style, e.g. mechanics.