ArM.Types.TraitKey

Description

We use TraitKey objects to reference arbitrary traits in a variety of contexts. It is conceptually simple, but the ordering takes some coding to handle all the different kinds of traits. To avoid cluttering, it is kept as a separate module.

JSON support is provided, where most traits are represented as an object of the form { kind : trait } or `{ "art" : Creo }`. Only a few traits require additional information to disambiguate.

Documentation

data TraitKey

A unique identifier for traits. It implement the Ord class, with canonical ordering on characteristics and arts.

Constructors

AbilityKey String

CharacteristicKey String

ArtKey String

SpellKey String Int String

PTraitKey String

ReputationKey String String

VFKey String String

ConfidenceKey String

OtherTraitKey String

SpecialKey String

PossessionKey String

CombatKey String

AgeKey

NoTrait