ArM.Char.Character

Description

This module contains exports the types to process characters and advancement, including persistence in JSON and advancement.

Documentation

module ArM.Char.Types.Character

module ArM.Types.KeyPair

module ArM.Char.Types.Advancement

data Advancement

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types.

Constructors

Advancement

Fields

advMode :: Maybe String

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: Maybe String

freeform description of the activities

advUses :: Maybe [Resource]

Books and other resources used exclusively by the character

advSQ :: Maybe XPType

Source Quality (SQ) This should be the common SQ for adventures; individual variation should be recorded as advBonus

advBonus :: Maybe XPType

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

prepareCharacter :: Character -> Character

Char Gen

Compute the initial state if no state is recorded.

The function uses applyCGA to process all of the pregame advancements. It then calls addConfidence to add the confidence trait to the state for the returned Character object

class Advance a where

Character Advancement

The Advance class represents object which change state from season to season.

Methods

advance :: SeasonTime -> a -> a

Advance the character until after the given time.

step :: a -> a

Advance the character one season forward

nextSeason :: a -> SeasonTime

Time of the next advancement of the character.

data Season

Season of the year. ``` data Season = Spring | Summer | Autumn | Winter | NoSeason ```

Constructors

Winter

Spring

Summer

Autumn

NoSeason

data SeasonTime

Generic definitions

A SeasonTime is a point in the narrative time, or NoTime for undefined.

Constructors

SeasonTime Season Int

GameStart

NoTime

characterEntryTime :: Character -> SeasonTime

Convenience Functions for Character Properties

The first season the character is played

isGameStart :: Character -> Bool

The first season the character is played

class HarmObject h where

The Harm Object

The HarmObject class establishes a common interface for Covenant and Character.

Minimal complete definition

name, stateSeason

Methods

name :: h -> String

Full name of the entity

stateSeason :: h -> SeasonTime

Current season of the object's stateY

stateName :: h -> String

String identifying the object and its state

prepare :: h -> h

The prepare function is applied when the object is read from file