ArM.Char.Trait

Description

Traits are represented as ProtoTrait objects in the design and advancement. A ProtoTrait can represent either a new trait or a change to an existing trait.

Processing advancements, the ProtoTrait is converted to a Trait object, representing the current state of the trait.

There are individual types for each kind of trait, e.g. virtue/flaw, ability, spell, art, etc. Conversion between a Trait object and a more specific type, like Art, is done by the (polymorphic) toTrait and getTrait functions.

Documentation

module ArM.Char.Types.Trait

data ProtoTrait

ProtoTrait

A ProtoTrait represents a new trait or an advancement of an existing trait as represented in the JSON input. Most fields are used only for certain types of traits and are Nothing for other types. This is the case, in particular, for the first quite a few fields which give the name of the trait of the relevant type only.

Constructors

ProtoTrait

Fields

ability :: Maybe String

ability name

virtue :: Maybe String

virtue name

flaw :: Maybe String

flaw name

characteristic :: Maybe String

characteristic name

art :: Maybe String

art name

spell :: Maybe String

spell name

ptrait :: Maybe String

personality trait name

confidence :: Maybe String

confidence, true faith, or similar

reputation :: Maybe String

reputation contents

other :: Maybe String

other trait, e.g. warping or decrepitude

strait :: Maybe String

special trait, like Longevity Potion

aging :: Maybe Aging

Aging object

possession :: Maybe Possession

Possesion includes weapon, vis, equipment, etc.

combat :: Maybe CombatOption

Possesion includes weapon, vis, equipment, etc.

protoTraitKey :: TraitKey

spec :: Maybe String

specialisation of an ability

detail :: Maybe String

detail (options) for a virtue or flaw

appliesTo :: Maybe TraitKey

not used (intended for virtues/flaws applying to another trait)

levelCap :: Maybe Int

cap on advancement

level :: Maybe Int

level of a spell

tefo :: Maybe String

technique/form of a spell

locale :: Maybe String

locale/domain of a reputation

mastery :: Maybe [String]

mastery options for a spell

flawless :: Maybe Bool

for a spell, if flawless magic applies

score :: Maybe Int

new score to replace the old one

bonusScore :: Maybe Int

bonus from puissant; should also be used on initial characteristics for the bonus from Strong Faerie Blood or Great/Poor Characteristic to keep it separate from the levels bought from points.

multiplyXP :: Maybe Float

XP multiplier from affinities and similar

cost :: Maybe Int

cost of a virtue or flaw

points :: Maybe Int

points for confidencetrue faithetc (additive)

xp :: Maybe XPType

XP to be added to the trait

agingPts :: Maybe Int

aging points for characteristicds (additive)

charBonuses :: [(Int, Int)]

bonuses from virtues to apply to a characteristic

multiplicity :: Maybe Int

number of types a virtue/flaw is taken; could be negative to remove an existing, but this is not yet implemented

comment :: Maybe String

freeform comment

data TraitKey

A unique identifier for traits. It implement the Ord class, with canonical ordering on characteristics and arts.

Constructors

AbilityKey String

CharacteristicKey String

ArtKey String

SpellKey String Int String

PTraitKey String

ReputationKey String String

VFKey String String

ConfidenceKey String

OtherTraitKey String

SpecialKey String

PossessionKey String

CombatKey String

AgeKey

NoTrait

class TraitType t where

Advancement - the TraitType class

Advancement of traits is based ProtoTrait objects representing changes in objects. The TraitType class provides two functions. Firstly advanceTrait applies a ProtoTrait to a Trait to advance it to a new Trait. Secondly, computeTrait creates a previously non-existing trait from a ProtoTrait.

These functions are implemented for each type representing a kind of trait.

The TraitType class provides the methods to advance traits using ProtoTrait objects.

Minimal complete definition

computeTrait

Methods

computeTrait :: ProtoTrait -> Maybe t

Convert a ProtoTrait (advancement) to a new trait object.

advanceTrait :: ProtoTrait -> t -> t

Advance a trait using changes defined by a ProtoTrait.

advanceTraitList :: [ProtoTrait] -> [Trait] -> [Trait]

Advancement

Apply a list of ProtoTrait advancements to a list of Traits.

This is the main function used by other modules when characters are advanced.

defaultPT :: ProtoTrait

Default ProtoTrait object, used internally for step-by-step construction of new objects.

spellKeyName :: TraitKey -> String

Convenience functions

Get the spell name from a TraitKey object

data Aging

Constructors

Aging

Fields

addYears :: Maybe Int

deltaYounger :: Maybe Int

Should be 1 when age changes and apparent age does not, otherwise 0

agingRollDie :: Maybe Int

aging roll die result

agingRoll :: Maybe Int

aging roll total

longevity :: Maybe Int

score of new longevity ritual

agingLimit :: Maybe Int

freeform comment

agingBonus :: Maybe Int

Bonus to aging rolls (excluding LR)

agingComment :: Maybe String

age when aging rolls are required

defaultAging :: Aging

Aging

data Weapon

Constructors

Weapon

Fields

weaponName :: String

weaponAbility :: String

weaponInit :: Int

atk :: Maybe Int

def :: Maybe Int

dam :: Maybe Int

str :: Maybe Int

range :: Maybe Int

load :: Int

weaponCost :: String

data Armour

Constructors

Armour

Fields

armourName :: String

armourLoad :: Int

armourProtection :: Int

armourCost :: String

findTrait :: TraitClass a => TraitKey -> [a] -> Maybe a

Find a trait, given by a key, from a list of Trait objects.

processChar :: Trait -> Trait

Postprocessing of traits

Process bonuses to characteristics, typically granted by virtues. THis has to be applied after validation of the points expenditure at char gen.