ArM.Char.Types.Advancement

Description

Documentation

(>*) :: Ord a => a -> a -> Bool

Calendar

The software assumes the Hibernian calendar, with Winter being the first season of the year. Several things may have to change for the standard calendar with Winter as the last season of the year. We have tried to collect these definitions here.

Hibernian Calendar

This comparison is used to check if a character is past the age limit where aging rolls are required. ``` (*>) = (>=) ```

data Season

Season of the year. ``` data Season = Spring | Summer | Autumn | Winter | NoSeason ```

Constructors

Winter

Spring

Summer

Autumn

NoSeason

data SeasonTime

Generic definitions

A SeasonTime is a point in the narrative time, or NoTime for undefined.

Constructors

SeasonTime Season Int

GameStart

NoTime

isWinter :: SeasonTime -> Bool

Is the Season Winter?

parseST :: String -> SeasonTime

Parse SeasonTime from String

parseSeasonTime :: Maybe String -> SeasonTime

Parse SeasonTime from `Maybe String`

data AdvancementType

Advancement

Constructors

Practice

Exposure

Adventure

Teaching

Training

Reading

VisStudy

data ExposureType

Constructors

LabWork

Teach

Train

Writing

Copying

OtherExposure

NoExposure

data Resource

Constructors

Resource String

class AdvancementLike a where

Methods

mode

Arguments

:: a

-> Maybe String
mode of study

season

Arguments

:: a

-> SeasonTime
season or development stage

narrative

Arguments

:: a

-> Maybe String
freeform description of the activities

uses

Arguments

:: a

-> Maybe [Resource]
Books and other resources used exclusively by the character

sourceQuality

Arguments

:: a

-> Maybe XPType
Source Quality (SQ) effectiveSQ :: Maybe Int -- ^ SQ modified by virtues and flaws

changes

Arguments

:: a

-> [ProtoTrait]
trait changes defined by player inferredTraits :: [ ProtoTrait ] -- ^ trait changes inferred by virtues and flaws

data Advancement

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types.

Constructors

Advancement

Fields

advMode :: Maybe String

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: Maybe String

freeform description of the activities

advUses :: Maybe [Resource]

Books and other resources used exclusively by the character

advSQ :: Maybe XPType

Source Quality (SQ) This should be the common SQ for adventures; individual variation should be recorded as advBonus

advBonus :: Maybe XPType

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

defaultAdv :: Advancement

data Validation

Constructors

ValidationError String

Validated String

data AugmentedAdvancement

Advancement with additional inferred fields

Constructors

Adv

Fields

advancement :: Advancement

Base advancement as entered by the user

effectiveSQ :: Maybe XPType

SQ modified by virtues and flaws

levelLimit :: Maybe Int

spell level allowance

spentXP :: Maybe XPType

Total XP spent on advancement

inferredTraits :: [ProtoTrait]

trait changes inferred by virtues and flaws

augYears :: Maybe Int

number of years advanced

validation :: [Validation]

Report from validation

postProcessTrait :: PostProcessor

extra postprocessing for traits at a given stage

data PostProcessor

Constructors

PostProcessor (Trait -> Trait)

defaultAA :: AugmentedAdvancement