ArM.Char.Types.Advancement

Description

Documentation

data AdvancementType

Advancement

Advancement Types

Main advancement modes as defined in the core rules [ArM]

Constructors

Practice

Adventure

Taught

Trained

Reading

VisStudy

Exposure ExposureType

CharGen String

data ExposureType

Different activities that grant Exposure but may require different processing in advancement and validation.

Constructors

LabWork

Teaching

Training

Writing

Copying

Authoring

Initiation

OpeningArts

Work

SpellInstruction

OtherExposure String

parseET :: String -> ExposureType

Parse an ExposureType from a String, auxiliary for FromJSON

parseAT :: String -> AdvancementType

Parse an AdvancementType from a String, auxiliary for FromJSON

dropWord :: String -> String

Drop the first word from a string.

class AdvancementLike a where

AdvancementLike Class

The AdvancementLike class gives a common API to Advancement and AugmentedAdvanceemnt

Minimal complete definition

mode, narrative, usesBook, sourceQuality, changes

Methods

mode

Arguments

:: a

-> AdvancementType
mode of study

narrative

Arguments

:: a

-> Maybe String
freeform description of the activities

usesBook

Arguments

:: a

-> [String]
Books used exclusively by the character

sourceQuality

Arguments

:: a

-> Maybe XPType
Source Quality (SQ)

changes

Arguments

:: a

-> [ProtoTrait]
trait changes defined by player

isExposure :: a -> Bool

data Advancement

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types. Note that standard SQ should be recorded as advSQ, while individual variation may be recorded as advBonus.

Constructors

Advancement

Fields

advMode :: AdvancementType

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: Maybe String

freeform description of the activities

advUses :: [String]

Books used exclusively by the character

advSQ :: Maybe XPType

Source Quality (SQ)

advBonus :: Maybe XPType

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

defaultAdv :: Advancement

data AugmentedAdvancement

The Augmented Advancement

Advancement with additional inferred fields

Constructors

Adv

Fields

advancement :: Advancement

Base advancement as entered by the user

baseSQ :: Maybe XPType

Base Source Quality

bonusSQ :: XPType

Bonus to Source Quality from Virtues and Flaws

levelLimit :: Maybe Int

spell level allowance

spentXP :: Maybe XPType

Total XP spent on advancement

inferredTraits :: [ProtoTrait]

trait changes inferred by virtues and flaws

augYears :: Maybe Int

number of years advanced

validation :: [Validation]

Report from validation

postProcessTrait :: PostProcessor

Extra postprocessing for traits at the given stage

bookUsed :: [Book]

Books required by the activity

teacherSQ :: Maybe Int

data PostProcessor

Type of function used to post-process traits after advancement.

Constructors

PostProcessor (Trait -> Trait)

defaultAA :: AugmentedAdvancement

effectiveSQ :: AugmentedAdvancement -> XPType

data Validation

Validation

A Validation is a message reporting either an error or a successful test.

Constructors

ValidationError String

Validated String