ArM.Char.Validation

Description

Only one function is exported: validate. It checks the integrity of an AugmentedAdvancement object, and adds its reports to the validation field.

This function should be used after all other calculation of the AugmentedAdvancement is complete, and before the object is displayed, or written to file.

Documentation

validate :: AugmentedAdvancement -> AugmentedAdvancement

validate an advancement, adding results to the validation field

validateXP :: AugmentedAdvancement -> AugmentedAdvancement

Validate allocation of XP.