ArM.Cov.Covenant

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data Covenant

Covenant Object

A Covenant consists of a state and a timeless concept, as well as lists of advancement which define the evolution of states

Constructors

Covenant

Fields

covenantConcept :: CovenantConcept

covenantState :: Maybe CovenantState

pastCovAdvancement :: [CovAdvancement]

futureCovAdvancement :: [CovAdvancement]

data CovenantID

ID of a Covenant. This is currently implemented as the name. It is used to reference the covenant, independtly of state, from other objects.

Constructors

CovenantID String

covenantID :: Covenant -> CovenantID

get the ID of a character.

data CovenantConcept

CovenantConcept Object

Constructors

CovenantConcept

Fields

covName :: String

covConcept :: Maybe String

covFounded :: Maybe Int

covAppearance :: Maybe String

covTribunal :: Maybe String

covData :: KeyPairList

defaultCovConcept :: CovenantConcept

Default (empty) covenant concept object.

data CovenantState

CovenantState Object

Constructors

CovenantState

Fields

covTime :: SeasonTime

covenFolkID :: [CharacterID]

covenFolk :: [Character]

library :: [BookCopy]

data CovAdvancement

Advancement and Traits

Advancement (changes) to a covenant.

Constructors

CovAdvancement

Fields

caSeason :: SeasonTime

season or development stage

caNarrative :: Maybe String

freeform description of the activities

joining :: [CharacterID]

leaving :: [CharacterID]

acquired :: [BookCopy]

lost :: [BookCopy]

defaultAdv :: CovAdvancement