ArM.Cov.Saga

Description

Documentation

data Saga

Saga type

A Saga as it is processed in memory. Multiple files have to be loaded to generate a Saga object from a SagaFile.

Constructors

Saga

Fields

sagaTitle :: String

sagaStates :: [SagaState]

rootDir :: String

sagaDesc :: String

baseURL :: Maybe String

spells :: SpellDB

weapons :: WeaponDB

armour :: ArmourDB

data SagaFile

SagaFile object

A Saga as it is stored on file. The main purpose here is to identify all the files used for characters and other data in the saga.

Constructors

SagaFile

Fields

title :: String

seasons :: [SeasonTime]

currentSeason :: SeasonTime

rootDirectory :: Maybe String

sagaDescription :: String

covenantFiles :: [String]

characterFiles :: [String]

spellFile :: String

weaponFile :: String

armourFile :: String

data SagaState

Saga state at a particular point in time, comprising characters and covenants at that point.

Constructors

SagaState

Fields

stateTitle :: String

seasonTime :: SeasonTime

covenants :: [Covenant]

characters :: [Character]

advanceSaga :: SagaFile -> Saga -> Saga

Advance the Saga according to timestamp in the SagaFile.

sagaStateName :: SagaState -> String

Get the name of the Saga as recorded in the SagaState

characterIndex :: [Character] -> OList

Write a bullet list of links for a list of characters

covenantIndex :: [Covenant] -> OList

Write a bullet list of links for a list of characters

advancementErrors :: SagaState -> OList

Error reports

Get an OList of all error messages from past advancements in a saga state.

CharGen errors are only included at GameStart and ignored later.

advancementErrorsLimit :: SeasonTime -> SagaState -> OList

Exctract a list of validation errors after a given time This is not currently used, but could be used to ignore old errors when reporting recent character states.