ArM.GameRules

Description

Simple and generic functions for ArM Rules. Should be independent of the software design.

Documentation

type XPType = Float

The XPType is the taype of XP. By RAW it should be Int, but some troupes prefer to count fractional XP when affinities are involved. This requires changing to a floating point type.

pyramidScore :: Num a => Int -> a

Calculate the XP cost for a given score on the pyramid scale

class GenericXPType a where

The class provides functions which must be implemented differently depending on the definition of XPType.

Methods

getAbilityScore :: Maybe a -> (Int, a)

scoreFromXP :: a -> Int

Calculate score from total XP, using the arts scale. For abilities, the argument should be divided by 5 beforehand.

calcXP :: Float -> a -> Maybe a -> a

xpround :: Float -> a

Round the XPType if required. This has to be redefined depending on the type of XPType.