ArM.Helper

Description

Documentation

trim :: String -> String

Trim away whitespace from the head and tail of the string.

(//) :: Integral a => a -> a -> a

Division of integers, rounding up

(-=) :: Ord a => [a] -> [a] -> [a]

Remove elements from the second list from the first. Both lists have to be sorted.

showstat :: Show a => Maybe a -> String

Convenience functions for Maybe

Show a number or «N/A»

maybeHead :: [a] -> Maybe a

return the head of a list or Nothing if the list is empty

maybeList :: Maybe [a] -> [a]

Get a list from a maybe-list, mapping Nothing to the empty list.

maybeAdd :: Num a => Maybe a -> Maybe a -> Maybe a

Add maybe-integers, treating Nothing as zero.

filterNothing :: [Maybe a] -> [a]

Remove Nothing elements from a list, and map the Just-elements to the constituent object.

passMaybe :: Monad m => (a -> m b) -> Maybe a -> m (Maybe b)

Skip Nothing values and apply the give function on Just-objects.

maybeShow :: Show a => Maybe a -> String

Convenience display functions

showNum :: (Show a, RealFrac a) => a -> String

Show a number with decimals only if required.

showBonus :: Int -> String

Show a non-zero integer with sign, or an empty string

showSigned :: Int -> String

Show an integer with sign

showStrList :: [String] -> String

Show a list of strings without brackets and quotes.

uniqueSort :: (Ord a, Eq a) => [a] -> [a]

Unique Sort

Sort the list and remove duplicates.

markdownLink :: String -> String -> String

Convenience functions for Markdown

Set a markdown link, escaping spaces in the link.

pagesLink :: String -> String

Set a link for github pages

wikiLink :: String -> String

Set a wikilink

markdownDL :: String -> String -> OList

Set an item for a description list in markdown