ArM.IO

Description

The functions in this module generally take a filename (String) and returns a Maybe object by attempting to parse the relevant file.

Some objects, like the Spell Database, are stored in CSV, others in JSON, including Character and Covenant.

Documentation

readSaga

Arguments

:: String
Filename

-> IO (Maybe Saga)

Load the saga and its constituent objects from the given file. This calls both readSagaFile and loadSaga.

writeSaga :: Saga -> IO ()