ArM.Types.Advancement

Description

Advancement is tricky because there are many different modes of study and many special cases. The basic structure uses two main types: + Advancement is the changes defined by the user. + AugmentedAdvancement comprises the Advancement object and additional inferred changes.

The Source Quality (SQ) is controled by several fields. + AugmentedAdvancement baseSQ is the base source quality inferred from other traits and general rules. + Advancement sourceQuality allows the user to enter the basic SQ manually. This is required for Practice, and may be used to override the baseSQ. If it differs from baseSQ, a warning is issued. + AugmentedAdvancement bonusSQ includes modifications derived automatically from virtues and flaws, and other individual circumstances. + Advancement bonusQuality should be used for individual modifications that cannot be automatically inferred, such as correspondent and Study Bonus.

Also not the scoreCap or sourceCap that caps the score that can be acquired from the source. This applies to books, trainers, and teachers.

The Advancement object has a list of changes which is ProtoTrait objects modifying existing traits. Similarly, AugmentedAdvancement has inferredTraits for additional implied changes.

Documentation

data Advancement

The Advancement Type

The advancement object has two roles. It can hold the advancemet from one season or chargen stage, as specified by the user. It can also hold additional field inferred by virtues and flaws. One may consider splitting these two functions into two types. Note that standard SQ should be recorded as advSQ, while individual variation may be recorded as advBonus.

Constructors

Advancement

Fields

advMode :: AdvancementType

mode of study

advSeason :: SeasonTime

season or development stage

advYears :: Maybe Int

number of years advanced

advNarrative :: [String]

narrative description of the activities

advComment :: [String]

freeform description of the activities

advUses :: [String]

Books used exclusively by the character

advBook :: [Book]

Books used exclusively by the character

advSQ :: Maybe XPType

Source Quality (SQ)

advCap :: Maybe Int

Source Quality (SQ)

advBonus :: [BonusSQ]

Bonus to Source Quality (SQ)

advChanges :: [ProtoTrait]

trait changes defined by player

advSpellLevels :: Maybe Int

spell level allowance

advTeacherSQ :: Maybe XPType

The SQ generated as teacher

advValidation :: [Validation]

Report from validation

advPostprocessTrait :: PostProcessor

Extra postprocessing for traits at the given stage

defaultAdvancement :: Advancement

data AugmentedAdvancement

The Augmented Advancement

Advancement with additional inferred fields

Constructors

Adv

Fields

explicitAdv :: Advancement

inferredAdv :: Advancement

class StoryObject a => AdvancementLike a where

The AdvancementLike Class

The AdvancementLike class gives a common API to Advancement and AugmentedAdvanceemnt

Minimal complete definition

mode, years, usesBook, bookUsed, sourceQuality, sourceCap, bonusSQ, changes, spellLevels, teacherSQ, validation, postprocessTrait, sortAdvTraits, addValidation, addProtoTrait

Methods

mode

Arguments

:: a

-> AdvancementType
mode of study

years :: a -> Maybe Int

usesBook

Arguments

:: a

-> [String]
Books used exclusively by the character

bookUsed

Arguments

:: a

-> [Book]
Books used exclusively by the character

sourceQuality

Arguments

:: a

-> Maybe XPType
Source Quality (SQ)

sourceCap

Arguments

:: a

-> Maybe Int
Level cap from the source of learning

bonusSQ :: a -> [BonusSQ]

changes

Arguments

:: a

-> [ProtoTrait]
trait changes defined by player

spellLevels

Arguments

:: a

-> Maybe Int
spell level allowance

teacherSQ

Arguments

:: a

-> Maybe XPType
The SQ generated as teacher

validation

Arguments

:: a

-> [Validation]
Report from validation

postprocessTrait

Arguments

:: a

-> PostProcessor
Extra postprocessing for traits at the given stage

isExposure :: a -> Bool

totalBonusSQ :: a -> XPType

effectiveSQ :: a -> Maybe XPType

sortAdvTraits :: a -> a

Sort the list of trait changes

spentXP :: a -> XPType

Count regular XP (excluding reputation) spent in an Advancement

spentLevels :: a -> Int

Count spell levels from an Advancement

addValidation :: [Validation] -> a -> a

addProtoTrait :: [ProtoTrait] -> a -> a

data AdvancementType

Advancement

Advancement Types

Main advancement modes as defined in the core rules [ArM]

Constructors

Practice

Adventure

Taught

Trained

Reading

VisStudy

Exposure ExposureType

CharGen String

data Validation

Validation

A Validation is a message reporting either an error or a successful test.

Constructors

ValidationError String

Validated String

ValidationWarning String

data PostProcessor

Type of function used to post-process traits after advancement.

Constructors

PostProcessor (Trait -> Trait)

data BonusSQ

Constructors

BonusSQ

Fields

sourceBonus :: XPType

bonusSource :: String

validateXP :: AugmentedAdvancement -> AugmentedAdvancement

Validate allocation of XP.