ArM.Types.Calendar

Description

Documentation

data SeasonTime

Generic definitions

A SeasonTime is a point in the narrative time, or NoTime for undefined.

Constructors

SeasonTime Season Int

GameStart

NoTime

data Season

Season of the year. ``` data Season = Spring | Summer | Autumn | Winter | NoSeason ```

Constructors

Winter

Spring

Summer

Autumn

NoSeason

parseSeasonTime :: Maybe String -> SeasonTime

Parse SeasonTime from `Maybe String`

isWinter :: SeasonTime -> Bool

Is the Season Winter?

seasonNext :: SeasonTime -> SeasonTime

seasonPrev :: SeasonTime -> SeasonTime

(>*) :: Ord a => a -> a -> Bool

Calendar

The software assumes the Hibernian calendar, with Winter being the first season of the year. Several things may have to change for the standard calendar with Winter as the last season of the year. We have tried to collect these definitions here.

Hibernian Calendar

This comparison is used to check if a character is past the age limit where aging rolls are required. ``` (*>) = (>=) ```

class Timed a where

The Timed Class

Minimal complete definition

season

Methods

season

Arguments

:: a

-> SeasonTime
season or development stage

(<::) :: a -> a -> Bool

(>::) :: a -> a -> Bool

compareTimed :: a -> a -> Ordering

mergeByTime :: [a] -> [a] -> [a]

mergeTimed :: [[a]] -> [a]