ArM.Types.Library

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data BookStats

The stats of a book as required for advancement mechanics.

Constructors

BookStats

Fields

topic :: TraitKey

quality :: Int

bookLevel :: Maybe Int

data Book

A book may be an original manuscript or a copy.

Constructors

Book

Fields

bookID :: String

bookTitle :: String

bookStats :: [BookStats]

list of stats per topic covered

bookCreator :: String

Creator of the copy or manuscript

bookDate :: SeasonTime

Time the copy was made

copiedFrom :: Maybe Book

Book copied or Nothing for an original manuscript

bookLocation :: String

Location whre the book was written or copied

bookAnnotation :: String

Additional information in free text

bookLanguage :: Maybe String

Language of the book

bookCount :: Int

Number of copies

data BookKey

Type for the unique identifier of an original book

Constructors

BookKey String

originalBook :: Book -> Book

The original of a given book

originalID :: Book -> String

The ID used to avoid rereading of tractatus

originalDate :: Book -> SeasonTime

The original date the book was authored

originalAuthor :: Book -> String

The original author of a given book

originalTitle :: Book -> String

The original author of a given book

bookKey :: Book -> BookKey

Get the unique identifier of an original book

readStats :: String -> (Maybe Int, Int)

CSV

readMaybeInt :: String -> Int

readTopic :: String -> String -> TraitKey

readTopic' :: String -> String -> TraitKey

makeBookStats :: String -> String -> String -> BookStats