ArM.Types.Library

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data BookStats

The stats of a book as required for advancement mechanics.

Constructors

BookStats

Fields

topic :: TraitKey

quality :: Maybe Int

bookLevel :: Maybe Int

reread :: Int

data Book

A book may be an original manuscript or a copy.

Constructors

Book

Fields

bookID :: String

bookTitle :: String

bookStats :: [BookStats]

list of stats per topic covered

bookCreator :: String

Creator of the copy or manuscript

bookDate :: SeasonTime

Time the copy was made

antologyOf :: [Book]

The book is an antology of multiple books

copiedFrom :: Maybe String

Book copied or Nothing for an original manuscript

bookLocation :: Maybe String

Location whre the book was written or copied

bookAnnotation :: [String]

Additional information in free text

bookLanguage :: Maybe String

Language of the book

bookCount :: Int

Number of copies

originalTitle :: Book -> String

The original author of a given book

originalAuthor :: Book -> String

The original author of a given book

originalDate :: Book -> SeasonTime

The original date the book was authored

originalID :: Book -> String

The ID used to avoid rereading of tractatus

bookKey :: Book -> HarmKey

Get the unique identifier of an original book

readStats :: String -> (Maybe Int, Maybe Int)

CSV

readMaybeInt :: String -> Int

readTopic :: String -> String -> TraitKey

readTopic' :: String -> String -> TraitKey

makeBookStats :: String -> String -> String -> BookStats

data ReadingID

Descriptions of a reading season

Constructors

ReadingID

Fields

bookRead :: HarmKey

partRead :: Maybe HarmKey

topicRead :: TraitKey