ArM.Types.Library

Description

This module contains types to process characters, including persistence in JSON and advancement.

Documentation

data BookStats

The stats of a book as required for advancement mechanics.

Constructors

BookStats

Fields

topic :: TraitKey

quality :: Int

bookLevel :: Maybe Int

data BookOriginal

An original manuscript with title, author, and stats.

Constructors

BookOriginal

Fields

bookTitle :: String

bookStats :: [BookStats]

bookAuthor :: String

bookYear :: Int

bookLocation :: String

bookAnnotation :: String

data BookKey

Type for the unique identifier of an original book

Constructors

BookKey String

bookKey :: BookOriginal -> BookKey

Get the unique identifier of an original book

data BookCopy

A copy of a book, whose stats may or may not deviate from the original manuscript.

Constructors

BookCopy

Fields

copyTitle :: String

copyStats :: [BookStats]

copyist :: String

copyYear :: Int

copyLocation :: String

copyAnnotation :: String

original :: BookKey